requirements.txt
# Function dependencies, for example:
google-cloud-compute==1.6.1
pytz==2018.5